+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Trade policy validation to user not approved.
+
 ## [0.4.1] - 2020-07-06
 ### Fixed